Moved to Cloudflare time server
Removed reliance on own timing server and replaced it with Cloudflare ones (Cloudflare will choose automatically its nearest server) and other performance improvements to syncClock() function. I was able to get roundtripdelay as low as 12.5 ms. This may not necessarily make a direct impact on the quality of the sync, but it will allow for future development in this direction.